Struct reth_tracing::RethTracer

pub struct RethTracer { /* private fields */ }

Tracer for application logging.

Manages the configuration and initialization of logging layers, including standard output, optional journald, and optional file logging.

Implementations

impl RethTracer

pub fn new() -> Self

Constructs a new Tracer with default settings.

Initializes with default stdout layer configuration. Journald and file layers are not set by default.

pub fn with_stdout(self, config: LayerInfo) -> Self

Sets a custom configuration for the stdout layer.

Arguments

config - The LayerInfo to use for the stdout layer.

pub fn with_journald(self, filter: String) -> Self

Sets the journald layer filter.

Arguments

filter - The filter to use for the journald layer.

pub fn with_file(self, config: LayerInfo, file_info: FileInfo) -> Self

Sets the file layer configuration and associated file info.

Arguments

config - The LayerInfo to use for the file layer.

file_info - The FileInfo containing details about the log file.